requirements.txt
celery==5.4.0
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
drf-extra-fields==3.7.0
factory_boy==3.3.1
filetype==1.2.0
gunicorn==23.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
requests==2.32.3